+++ TravelingSalesmanProblem.h
@@ -63,11 +63,13 @@ Tour* Christofides_FindTour(const Graph* g);
 Tour* TwoOpt_ImproveTour(const Graph* g, Tour* initialTour);
 Tour* OrOpt_ImproveTour(const Graph* g, Tour* initialTour);
 Tour* ThreeOpt_ImproveTour(const Graph* g, Tour* initialTour);
+Tour* LinKernighan_ImproveTour(const Graph* g, Tour* initialTour);
 
 // Metaheuristics (Implemented in their respective .c files)
 Tour* SimulatedAnnealing_FindTour(const Graph* g, unsigned int* initialTour);
 Tour* AntColony_FindTour(const Graph* g);
 Tour* GeneticAlgorithm_FindTour(const Graph* g);
+Tour* TabuSearch_FindTour(const Graph* g);
 
 // In the future, it could be interesting implementing ISPO:
 // https://www.researchgate.net/publication/271285365_ISPO_A_New_Way_to_Solve_Traveling_Salesman_Problem